+++ test/main.test.js
@@ -7,8 +7,10 @@ var config = {
   TaskDefinition: 'arn:aws:ecs:us-east-1:123456789012:task-definition/fake:1',
   Concurrency: '3',
   QueueUrl: 'https://fake.us-east-1/sqs/url',
+  TaskEventQueueUrl: 'https://fake.us-east-1/sqs/url-for-events',
   StackName: 'watchbot-testing',
   ExponentialBackoff: false
+  // , LogLevel: 'debug'
 };
 
 util.mock('[main] message polling error', function(assert) {
@@ -20,7 +22,6 @@ util.mock('[main] message polling error', function(assert) {
 
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 0, 'no ecs.describeTasks requests');
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     var errorMsg = context.logs.find(function(log) {
       return /Mock SQS error/.test(log);
@@ -41,7 +42,6 @@ util.mock('[main] nothing to do', function(assert) {
 
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 0, 'no ecs.describeTasks requests');
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.end();
   });
@@ -56,7 +56,6 @@ util.mock('[main] run a task', function(assert) {
 
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 1, 'one ecs.describeTasks requests');
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.deepEqual(context.ecs.runTask, [
       {
@@ -92,7 +91,6 @@ util.mock('[main] task running error', function(assert) {
 
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 0, 'no ecs.describeTasks requests');
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.equal(context.ecs.runTask.length, 1, '1 ecs.runTask request');
     assert.ok(context.logs.find(function(log) {
@@ -121,10 +119,9 @@ util.mock('[main] task running failure (out of memory)', function(assert) {
 
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 0, 'no ecs.describeTasks requests');
-    assert.equal(context.sqs.receiveMessage.length, 2, 'one sqs.receiveMessage requests');
+    assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.equal(context.ecs.runTask.length, 1, '1 ecs.runTask request');
-    util.collectionsEqual(assert, context.sns.publish, [], 'doesn\'t send error notification');
+    assert.deepEqual(context.sns.publish, [], 'does not send failure notification');
     util.collectionsEqual(assert, context.sqs.changeMessageVisibility, [
       { ReceiptHandle: '1', VisibilityTimeout: 0 }
     ], 'expected sqs.changeMessageVisibility requests');
@@ -141,7 +138,6 @@ util.mock('[main] task running failure (unrecognized reason)', function(assert) 
 
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 0, 'no ecs.describeTasks requests');
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.equal(context.ecs.runTask.length, 1, '1 ecs.runTask request');
     assert.deepEqual(context.sns.publish, [], 'does not send failure notification');
@@ -161,7 +157,6 @@ util.mock('[main] message completion error after task run failure', function(ass
 
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 0, 'no ecs.describeTasks requests');
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.equal(context.ecs.runTask.length, 1, '1 ecs.runTask request');
     assert.ok(context.logs.find(function(log) {
@@ -187,20 +182,18 @@ util.mock('[main] task polling error', function(assert) {
     { MessageId: 'task-failure', ReceiptHandle: '1', Body: JSON.stringify({ Subject: 'subject1', Message: 'message1' }), Attributes: { SentTimestamp: 10, ApproximateReceiveCount: 0, ApproximateFirstReceiveTimestamp: 20 } }
   ];
 
+  context.sqs.eventMessages = [
+    { MessageId: 'error', Attributes: { ApproximateReceiveCount: 0 } }
+  ];
+
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 1, 'one ecs.describeTasks requests');
-    assert.equal(context.sqs.receiveMessage.length, 1, 'one sqs.receiveMessage requests');
+    assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.equal(context.ecs.runTask.length, 1, '1 ecs.runTask request');
     assert.ok(context.logs.find(function(log) {
-      return /Mock ECS error/.test(log);
+      return /Mock SQS error/.test(log);
     }), 'printed error message');
-    assert.deepEqual(context.sns.publish, [
-      {
-        Subject: '[watchbot] task polling error',
-        Message: 'Mock ECS error'
-      }
-    ], 'sent expected error notification');
+    assert.deepEqual(context.sns.publish, [], 'sent no error notification');
     assert.end();
   });
 });
@@ -219,15 +212,6 @@ util.mock('[main] no free tasks', function(assert) {
   watchbot.main(config).on('finish', function() {
     assert.equal(context.sqs.receiveMessage.length, 1, 'one sqs.receiveMessage requests');
     assert.equal(context.ecs.runTask.length, 3, 'three ecs.runTask requests');
-    util.collectionsEqual(assert, context.ecs.describeTasks, [
-      {
-        tasks: [
-          '3b80fe64b7d8278090a63a16e5908ad9',
-          '530ca7ee1cdf371158f9eeba094ae113',
-          '6599a09f3128ed099d8eafff0b832728'
-        ]
-      }
-    ], 'expected ecs.describeTasks requests');
     assert.end();
   });
 });
@@ -243,6 +227,43 @@ util.mock('[main] manage messages for completed tasks', function(assert) {
     { MessageId: 'finish-4', ReceiptHandle: '4', Body: JSON.stringify({ Subject: 'subject4', Message: 'message4' }), Attributes: { SentTimestamp: 10, ApproximateReceiveCount: 0, ApproximateFirstReceiveTimestamp: 20 } }
   ];
 
+  context.sqs.eventMessages = context.sqs.messages.map((message) => {
+    return {
+      MessageId: `${message.MessageId}-event`,
+      ReceiptHandle: `${message.ReceiptHandle}-event`,
+      Attributes: {
+        SentTimestamp: message.Attributes.SentTimestamp,
+        ApproximateReceiveCount: 0
+      },
+      Body: JSON.stringify({
+        detail: {
+          clusterArn: 'cluster-arn',
+          containerInstanceArn: 'instance-arn',
+          taskArn: util.expectedArn(message, config.TaskDefinition, config.ContainerName, config.StackName),
+          lastStatus: 'STOPPED',
+          stoppedReason: message.MessageId.split('-')[1],
+          overrides: {
+            containerOverrides: [
+              {
+                environment: [
+                  { name: 'MessageId', value: message.MessageId },
+                  { name: 'Subject', value: JSON.parse(message.Body).Subject },
+                  { name: 'Message', value: JSON.parse(message.Body).Message },
+                  { name: 'SentTimestamp', value: message.Attributes.SentTimestamp },
+                  { name: 'ApproximateFirstReceiveTimestamp', value: message.Attributes.ApproximateFirstReceiveTimestamp },
+                  { name: 'ApproximateReceiveCount', value: message.Attributes.ApproximateReceiveCount + 1 }
+                ]
+              }
+            ]
+          },
+          containers: [{ exitCode: Number(message.MessageId.split('-')[1]) }],
+          startedAt: 1484155849718,
+          stoppedAt: 1484155857691
+        }
+      })
+    };
+  });
+
   var testConfig = Object.assign({}, config, {
     Concurrency: '5',
     NotifyAfterRetries: '1'
@@ -250,13 +271,17 @@ util.mock('[main] manage messages for completed tasks', function(assert) {
 
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(testConfig).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 1, 'one ecs.describeTasks request');
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.equal(context.ecs.runTask.length, 5, 'five ecs.runTask requests');
     util.collectionsEqual(assert, context.sqs.deleteMessage, [
+      { ReceiptHandle: '0-event' },
+      { ReceiptHandle: '1-event' },
+      { ReceiptHandle: '2-event' },
+      { ReceiptHandle: '3-event' },
+      { ReceiptHandle: '4-event' },
       { ReceiptHandle: '0' },
       { ReceiptHandle: '3' }
-    ], 'expected sqs.deleteMessage requests');
+    ], ' sqs.deleteMessage for all event messages, and for expected job messages');
     util.collectionsEqual(assert, context.sqs.changeMessageVisibility, [
       { ReceiptHandle: '1', VisibilityTimeout: 0 },
       { ReceiptHandle: '2', VisibilityTimeout: 0 },
@@ -284,9 +309,45 @@ util.mock('[main] message completion error', function(assert) {
     { MessageId: 'finish-0', ReceiptHandle: 'error', Body: JSON.stringify({ Subject: 'subject0', Message: 'message0' }), Attributes: { SentTimestamp: 10, ApproximateReceiveCount: 0, ApproximateFirstReceiveTimestamp: 20 } }
   ];
 
+  context.sqs.eventMessages = context.sqs.messages.map((message) => {
+    return {
+      MessageId: `${message.MessageId}-event`,
+      ReceiptHandle: `${message.ReceiptHandle}-event`,
+      Attributes: {
+        SentTimestamp: message.Attributes.SentTimestamp,
+        ApproximateReceiveCount: 0
+      },
+      Body: JSON.stringify({
+        detail: {
+          clusterArn: 'cluster-arn',
+          containerInstanceArn: 'instance-arn',
+          taskArn: util.expectedArn(message, config.TaskDefinition, config.ContainerName, config.StackName),
+          lastStatus: 'STOPPED',
+          stoppedReason: message.MessageId.split('-')[1],
+          overrides: {
+            containerOverrides: [
+              {
+                environment: [
+                  { name: 'MessageId', value: message.MessageId },
+                  { name: 'Subject', value: JSON.parse(message.Body).Subject },
+                  { name: 'Message', value: JSON.parse(message.Body).Message },
+                  { name: 'SentTimestamp', value: message.Attributes.SentTimestamp },
+                  { name: 'ApproximateFirstReceiveTimestamp', value: message.Attributes.ApproximateFirstReceiveTimestamp },
+                  { name: 'ApproximateReceiveCount', value: message.Attributes.ApproximateReceiveCount + 1 }
+                ]
+              }
+            ]
+          },
+          containers: [{ exitCode: Number(message.MessageId.split('-')[1]) }],
+          startedAt: 1484155849718,
+          stoppedAt: 1484155857691
+        }
+      })
+    };
+  });
+
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(config).on('finish', function() {
-    assert.equal(context.ecs.describeTasks.length, 1, 'one ecs.describeTasks request');
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
     assert.equal(context.ecs.runTask.length, 1, 'one ecs.runTask requests');
     assert.equal(context.sqs.changeMessageVisibility.length, 0, 'no sqs.changeMessageVisibility requests');
@@ -295,7 +356,7 @@ util.mock('[main] message completion error', function(assert) {
     });
     assert.ok(errorMsg, 'logged error');
     util.collectionsEqual(assert, context.sqs.deleteMessage, [
-      { ReceiptHandle: 'error' }
+      { ReceiptHandle: 'error' }, { ReceiptHandle: 'error-event' }
     ], 'expected sqs.deleteMessage request');
 
     util.collectionsEqual(assert, context.sns.publish, [